remove jenkins stuff
