--- .github/workflows/buildconfig/global.txt
@@ -1,124 +1,0 @@
-CMAKE_CXX_STANDARD=17
-alien=OFF
-arrow=OFF
-asimage=ON
-asserts=OFF
-builtin_afterimage=OFF
-builtin_cfitsio=OFF
-builtin_clang=ON
-builtin_cling=ON
-builtin_cppzmq=OFF
-builtin_davix=OFF
-builtin_fftw3=OFF
-builtin_freetype=OFF
-builtin_ftgl=OFF
-builtin_gl2ps=OFF
-builtin_glew=OFF
-builtin_gsl=OFF
-builtin_llvm=ON
-builtin_lz4=OFF
-builtin_lzma=OFF
-builtin_nlohmannjson=OFF
-builtin_openssl=OFF
-builtin_openui5=ON
-builtin_pcre=OFF
-builtin_tbb=OFF
-builtin_unuran=OFF
-builtin_vc=OFF
-builtin_vdt=OFF
-builtin_veccore=OFF
-builtin_xrootd=OFF
-builtin_xxhash=OFF
-builtin_zeromq=OFF
-builtin_zlib=OFF
-builtin_zstd=OFF
-ccache=OFF
-cefweb=OFF
-clad=ON
-cocoa=OFF
-coverage=OFF
-cuda=OFF
-cudnn=ON
-cxxmodules=OFF
-daos=OFF
-dataframe=ON
-davix=ON
-dcache=OFF
-dev=OFF
-distcc=OFF
-exceptions=ON
-fcgi=OFF
-fftw3=ON
-fitsio=ON
-fortran=OFF
-gdml=ON
-gfal=ON
-gnuinstall=OFF
-gsl_shared=OFF
-gviz=OFF
-http=ON
-imt=ON
-jemalloc=OFF
-libcxx=OFF
-macos_native=OFF
-mathmore=ON
-memory_termination=OFF
-minuit2=ON
-mlp=ON
-monalisa=OFF
-mpi=OFF
-mysql=ON
-odbc=OFF
-opengl=ON
-oracle=ON
-pgsql=ON
-pyroot_legacy=OFF
-pyroot=ON
-pythia6_nolink=OFF
-pythia6=ON
-pythia8=ON
-qt5web=OFF
-qt6web=OFF
-r=OFF
-roofit_hs3_ryml=OFF
-roofit_multiprocess=OFF
-roofit=ON
-root7=ON
-rpath=ON
-runtime_cxxmodules=ON
-shadowpw=OFF
-shared=ON
-soversion=OFF
-spectrum=ON
-sqlite=ON
-ssl=ON
-tcmalloc=OFF
-test_distrdf_dask=OFF
-test_distrdf_pyspark=OFF
-tmva-cpu=ON
-tmva-gpu=OFF
-tmva=ON
-tmva-pymva=ON
-tmva-rmva=OFF
-tmva-sofie=OFF
-unuran=OFF
-uring=OFF
-vc=OFF
-vdt=ON
-veccore=OFF
-vecgeom=OFF
-webgui=ON
-win_broken_tests=OFF
-winrtdebug=OFF
-x11=ON
-xml=ON
-xproofd=OFF
-xrootd=ON
-all=OFF
-asan=OFF
-clingtest=OFF
-gminimal=OFF
-minimal=OFF
-rootbench=OFF
-roottest=OFF
-testing=OFF
