--- .github/workflows/buildconfig/ubuntu20.txt
@@ -1,109 +1,0 @@
-alien=Off
-arrow=Off
-asimage=On
-asserts=ON
-builtin_afterimage=On
-builtin_cfitsio=Off
-builtin_cppzmq=On
-builtin_davix=On
-builtin_fftw3=Off
-builtin_freetype=Off
-builtin_ftgl=Off
-builtin_gl2ps=Off
-builtin_glew=Off
-builtin_gsl=Off
-builtin_gtest=On
-builtin_lz4=On
-builtin_lzma=Off
-builtin_nlohmannjson=On
-builtin_openssl=Off
-builtin_openui5=On
-builtin_pcre=Off
-builtin_tbb=On
-builtin_unuran=On
-builtin_vc=On
-builtin_vdt=On
-builtin_veccore=On
-builtin_xrootd=On
-builtin_xxhash=On
-builtin_zeromq=On
-builtin_zlib=Off
-builtin_zstd=Off
-ccache=ON
-cefweb=Off
-clad=On
-CMAKE_BUILD_TYPE=RelWithDebInfo
-CMAKE_CXX_STANDARD=17
-cocoa=Off
-cuda=Off
-cudnn=Off
-daos=Off
-dataframe=On
-davix=On
-dcache=Off
-dev=Off
-distcc=Off
-fail-on-missing=On
-fcgi=Off
-fftw3=On
-fitsio=On
-fortran=On
-gdml=On
-gfal=Off
-gsl_shared=Off
-gviz=On
-http=On
-imt=On
-libcxx=Off
-LLVM_ENABLE_ASSERTIONS=On
-mathmore=On
-minuit2=On
-minuit2_mpi=Off
-minuit2_omp=Off
-mlp=On
-monalisa=Off
-mpi=Off
-mysql=On
-odbc=On
-opengl=On
-oracle=Off
-pgsql=On
-pyroot2=Off
-pyroot3=On
-pyroot=On
-pyroot_legacy=Off
-pythia6=Off
-pythia6_nolink=Off
-pythia8=On
-qt5web=On
-qt6web=Off
-r=On
-roofit=On
-roofit_hs3_ryml=Off
-roofit_multiprocess=On
-roottest=ON
-runtime_cxxmodules=Off
-shadowpw=On
-spectrum=On
-sqlite=On
-ssl=On
-test_distrdf_dask=On
-test_distrdf_pyspark=On
-testing=ON
-tmva-cpu=On
-tmva-gpu=Off
-tmva-pymva=On
-tmva-rmva=On
-tmva-sofie=On
-tmva=On
-unuran=On
-uring=Off
-vc=On
-vdt=On
-veccore=On
-vecgeom=Off
-win_broken_tests=Off
-x11=On
-xml=On
-xproofd=Off
-xrootd=On
