+++ .github/workflows/rootci-installers/buildconfig/global.txt
@@ -1,6 +1,8 @@
 CMAKE_CXX_STANDARD=17
 alien=OFF
+all=OFF
 arrow=OFF
+asan=OFF
 asimage=ON
 asserts=OFF
 builtin_afterimage=OFF
@@ -35,6 +37,7 @@ builtin_zstd=OFF
 ccache=OFF
 cefweb=OFF
 clad=ON
+clingtest=OFF
 cocoa=OFF
 coverage=OFF
 cuda=OFF
@@ -53,6 +56,7 @@ fitsio=ON
 fortran=OFF
 gdml=ON
 gfal=ON
+gminimal=OFF
 gnuinstall=OFF
 gsl_shared=OFF
 gviz=OFF
@@ -63,6 +67,7 @@ libcxx=OFF
 macos_native=OFF
 mathmore=ON
 memory_termination=OFF
+minimal=OFF
 minuit2=ON
 mlp=ON
 monalisa=OFF
@@ -72,18 +77,20 @@ odbc=OFF
 opengl=ON
 oracle=ON
 pgsql=ON
-pyroot_legacy=OFF
 pyroot=ON
-pythia6_nolink=OFF
+pyroot_legacy=OFF
 pythia6=ON
+pythia6_nolink=OFF
 pythia8=ON
 qt5web=OFF
 qt6web=OFF
 r=OFF
+roofit=ON
 roofit_hs3_ryml=OFF
 roofit_multiprocess=OFF
-roofit=ON
 root7=ON
+rootbench=OFF
+roottest=OFF
 rpath=ON
 runtime_cxxmodules=ON
 shadowpw=OFF
@@ -95,12 +102,13 @@ ssl=ON
 tcmalloc=OFF
 test_distrdf_dask=OFF
 test_distrdf_pyspark=OFF
+testing=OFF
 tmva-cpu=ON
 tmva-gpu=OFF
-tmva=ON
 tmva-pymva=ON
 tmva-rmva=OFF
 tmva-sofie=OFF
+tmva=ON
 unuran=OFF
 uring=OFF
 vc=OFF
@@ -114,11 +122,3 @@ x11=ON
 xml=ON
 xproofd=OFF
 xrootd=ON
-all=OFF
-asan=OFF
-clingtest=OFF
-gminimal=OFF
-minimal=OFF
-rootbench=OFF
-roottest=OFF
-testing=OFF
