+++ .github/workflows/rootci-installers/buildconfig/ubuntu18.txt
@@ -1,3 +1,6 @@
+CMAKE_BUILD_TYPE=RelWithDebInfo
+CMAKE_CXX_STANDARD=17
+LLVM_ENABLE_ASSERTIONS=On
 alien=Off
 arrow=Off
 asimage=On
@@ -32,8 +35,6 @@ builtin_zstd=Off
 ccache=ON
 cefweb=Off
 clad=On
-CMAKE_BUILD_TYPE=RelWithDebInfo
-CMAKE_CXX_STANDARD=17
 cocoa=Off
 cuda=Off
 cudnn=Off
@@ -55,7 +56,6 @@ gviz=On
 http=On
 imt=On
 libcxx=Off
-LLVM_ENABLE_ASSERTIONS=On
 mathmore=On
 minuit2=On
 minuit2_mpi=Off
