--- .jenkins/s3/rc.sh
@@ -1,8 +1,0 @@
-
-export OS_AUTH_TYPE=v3applicationcredential
-export OS_AUTH_URL=https://keystone.cern.ch/v3
-export OS_IDENTITY_API_VERSION=3
-export OS_REGION_NAME="cern"
-export OS_INTERFACE=public
-#export OS_APPLICATION_CREDENTIAL_ID= # omitted
-# export OS_APPLICATION_CREDENTIAL_SECRET= # omitted
